+++ classes/event/observer.php
@@ -1,0 +1,49 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+namespace format_flexsections\event;
+
+/**
+ * Event observers for flexible sections course format.
+ *
+ * @package    format_flexsections
+ * @copyright  2022 Ruslan Kabalin
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class observer {
+
+    /**
+     * Triggered via \core\event\course_updated event.
+     *
+     * @param \core\event\course_updated $event
+     */
+    public static function course_updated(\core\event\course_updated $event) {
+        if (class_exists('format_flexsections', false)) {
+            \format_flexsections::update_end_date($event->courseid);
+        }
+    }
+
+    /**
+     * Triggered via \core\event\course_section_deleted event.
+     *
+     * @param \core\event\course_section_deleted $event
+     */
+    public static function course_section_deleted(\core\event\course_section_deleted $event) {
+        if (class_exists('format_flexsections', false)) {
+            \format_flexsections::update_end_date($event->courseid);
+        }
+    }
+}